Change how glk args are marshalled.
We want to be able to marshall arbitrary selections of args from
the glulx glk call into a single argument for the glk function, e.g.
a memptr and a length is marshalled into a string.

We also need to use arbitrary selections of args from the glulx glk
call and arbitrary selections of members of the result from the glk
code when calling a store function, e.g. the memptr and buffer length
from the glulx glk call, and the string result from the glk function
that was run, is used by the store function to save the string into
glulx memory.